refact: new Fortran bindings with test program

+++ src/test/test_matrix.cpp
@@ -43,12 +43,12 @@ int main (int argc, char *argv[])
     using namespace std;
     matrix m1(5, 6, true);
     srand(time(0));
-    for (int i = 0; i < m1.size; i++)
+    for (size_t i = 0; i < m1.size; i++)
         m1.c[i] = rand();
 
     matrix m2(m1);
     assert(m1.size == m2.size);
-    for (int i = 0; i < m1.size; i++)
+    for (size_t i = 0; i < m1.size; i++)
         if ( m1.c[i] != m2.c[i] ) throw logic_error("copy constructor fail");
     if (! (m1 == m2 )) throw logic_error("direct compare fail");
